% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{butland_ppi}
\alias{butland_ppi}
\title{Protein-protein interaction data.}
\format{A binary, symmetric matrix}
\usage{
data(butland_ppi)
}
\description{
A network of protein-protein interactions (bindings) from
Butland et al (2005) ``Interaction network containing conserved
    and essential protein complexes in Escherichia coli''.
}
\examples{
data(butland_ppi)
gofstats(butland_ppi)
}
\keyword{datasets}

