% Generated by roxygen2 (4.0.2): do not edit by hand
\name{casia}
\alias{casia}
\title{Central Asia Cooperation}
\format{A list including the following dyadic and nodal variables:
\itemize{
\item \code{coop}: a sociomatrix of counts of cooperative events;
\item \code{dist}: between-country distance;
\item \code{pop}: country population;
}}
\source{
Originally from the KEDS project, see
P.D. Hoff (2005) "Bilinear mixed-effects models for dyadic data" for
details.
}
\description{
Counts of between-country military cooperation in Central Asia
}

