% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_click.R
\docType{data}
\name{click}
\alias{click}
\title{Dataset of Web hits in chapter15}
\format{A data frame with 44 rows and 5 variables:
\describe{
  \item{click}{the quantity of web hits}
  \item{key_length}{the length of the keywords}
  \item{display}{display times}
  \item{price}{average price of web hits}
  \item{grade}{the average rank}
}}
\usage{
click
}
\description{
A dataset containing the length of keywords,display times,price of hits and average grade of 40 round objects.
}
\keyword{datasets}

