% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_consumption_income.R
\docType{data}
\name{consumption_income}
\alias{consumption_income}
\title{Dataset of income and consumption in chapter12}
\format{a dataframe with 36 rows and 3 variables
\describe{
 \item{Y}{year}
 \item{C}{consumption}
 \item{Y}{income}
}}
\usage{
consumption_income
}
\description{
A dataset containing year,consumption and income 2 variables of 36 objects
}
\keyword{datasets}

