% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CTG.R
\docType{data}
\name{CTG}
\alias{CTG}
\title{Dataset of fetal state}
\format{a dataframe with 2126 rows and 23 variables
\describe{
 \item{LB}{FHR basic line:heart rate per minute}
 \item{AC}{Accumulate times per second}
 \item{FM}{Fetal movement per second}
 \item{UC}{uterine contraction per second}
 \item{DL}{Light deceleration per second}
 \item{DS}{Serious deceleration per second}
 \item{DP}{Persistent deceleration per second}
 \item{ASTV}{Short term variation in the percentage of time}
 \item{MSTV}{Short term variation of average}
 \item{ALTV}{Long term variation in the percentage of time}
 \item{MLTCV}{Long term variation of average}
 \item{Width}{Width of FHR histogram}
 \item{Min}{The minimum value of FHR histogram}
 \item{Max}{The maximum value of FHR histogram}
 \item{Nmax}{The peak count of FHR histogram}
 \item{Nzeros}{The zero count of FHR histogram}
 \item{Mode}{The mode of FHR histogram}
 \item{Mean}{The mean of FHR histogram}
 \item{Median}{The median of FHR histogram}
 \item{Variance}{The variance of FHR histogram}
 \item{Tendency}{The tendency of FHR histogram}
 \item{CLASS}{The class of FHR}
 \item{NSP}{The class of fetal state}
}}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Cardiotocography}
}
\usage{
CTG
}
\description{
A dataset containing LB,AC,FM...23 variables of 2126 obejects
}
\keyword{datasets}

