% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_douglas.R
\docType{data}
\name{douglas}
\alias{douglas}
\title{Dataset of argiculture production in chapter12}
\format{a dataframe with 15 rows and 3 variables
\describe{
 \item{y}{actual output}
 \item{x2}{labor time}
 \item{x3}{actual capital input}
}}
\usage{
douglas
}
\description{
A dataset containing y,x2,x3 of 15 objects
}
\keyword{datasets}

