% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_EDA.R
\name{EDA}
\alias{EDA}
\title{Function of 'Exploratory Data Analysis'}
\usage{
EDA(x)
}
\arguments{
\item{x}{Numeric vectors}
}
\value{
The outputs will be several graphs to describe some features of our data.
}
\description{
Function of 'Exploratory Data Analysis'
}
\examples{
pay <- c(11,19,14,22,14,28,13,81,12,43,11,16,
31,16,23,42,22,26,17,22,13,27,108,16,43,82,
14,11,51,76,28,66,29,14,14,65,37,16,37,35,
39,27,14,17,13,38,28,40,85,32,25,26,16,12,
54,40,18,27,16,14,33,29,77,50,19,34)
EDA(pay)
log.pay <- log10(pay)
EDA(log.pay)

}

