% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fair.R
\docType{data}
\name{fair}
\alias{fair}
\title{Dataset of fairmodel study in chapter15}
\format{A data frame with 601 rows and 10 variables:
\describe{
 \item{ID}{the ID of objects}
 \item{affairs}{times of affairs,0:no;7:4~10times;12:more}
 \item{gender}{1:male;0:female}
 \item{age}{the age of objects}
 \item{yearsmarried}{married time}
 \item{children}{0:no;1:yes}
 \item{religiousness}{1~5:the degree of being religious upward}
 \item{education}{9:junior high school;12:high school;14,16,17,18,20:phd or others}
 \item{occupation}{1~7:by Hollingshead}
 \item{rating}{1~5:evaluation on marriage myself,from quite sad to very happy}
 }}
\source{
\url{http://farimodel.econ.yale.edu/rayfair/worksd.html}
}
\usage{
fair
}
\description{
A dataset containing affairs,gender,age...10 variables of 601 objects.
}
\keyword{datasets}

