% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fiscal_revenue.R
\docType{data}
\name{fiscal_revenue}
\alias{fiscal_revenue}
\title{Dataset of national revenue in chapter11}
\format{a dataframe with 15 rows and 8 variables
\describe{
 \item{t}{year}
 \item{industry}{industry output}
 \item{agriculture}{agriculture output}
 \item{construction}{construction output}
 \item{consumption}{social commodities value}
 \item{pop}{total population}
 \item{disaster}{disaster area}
 \item{revenue}{national revenue}
}}
\usage{
fiscal_revenue
}
\description{
A dataset containing t,industry,agriculture...8 variables of 15 objects
}
\keyword{datasets}

