% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfriends}
\alias{gfriends}
\title{Girls high school friendship data}
\format{A list including the following dyadic and nodal variables:
\itemize{
\item \code{Y}: an undirected sociomatrix of friendship ties;
\item \code{X}: nodal covariates including grade and normalized gpa and smoking scores
}}
\description{
Directed inditicaor of friendship between 144 high school girls
}

