% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Goldfeld_Quandt_test.R
\name{GQtest}
\alias{GQtest}
\title{Function of 'Goldfeld-Quandt test' for different variance}
\usage{
GQtest(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\value{
The outputs are P value and F value
}
\description{
Function of 'Goldfeld-Quandt test' for different variance
}
\examples{
data("plantarea_outputvalue")
GQtest(plantarea_outputvalue$plant_area,plantarea_outputvalue$output_value)


}

