% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grade.R
\docType{data}
\name{grade}
\alias{grade}
\title{Dataset of grade in chapter13}
\format{a dataframe with 32 rows and 4 variables
\describe{
 \item{GRADE}{1:grade upward;0:others}
 \item{GPA}{average grade}
 \item{TUCE}{economic grade in history}
 \item{PSI}{1:guided by new teaching method';0:others}
}}
\usage{
grade
}
\description{
A dataset containing GRADE,GPA,TUCE and PSI 4 variables of 32 objects
}
\keyword{datasets}

