% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_graduate_apply.R
\docType{data}
\name{graduate_apply}
\alias{graduate_apply}
\title{Dataset of graduate application in chapter14}
\format{a data frame with 4 variables of 400 objects
\describe{
 \item{apply}{willingness,0:unlikely;1:somewhat likely;2:very likely}
 \item{pared}{virtual variable for at least one parent is graduate,0:no;1:yes}
 \item{public}{virtual variable for ungraduate school,0:private;1:public}
 \item{gpa}{average grade}
}}
\source{
\url{http://www.ats.ucla.edu/stat/data/ologit.dta}
}
\usage{
graduate_apply
}
\description{
A dataset containing apply,pared,public and gpa of 400 objects
}
\keyword{datasets}

