% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grainproduct.R
\docType{data}
\name{grainproduct}
\alias{grainproduct}
\title{Dataset of grain production in chapter12}
\format{a dataframe with 27 rows and 5 variables
\describe{
 \item{year}{time}
 \item{Y}{output}
 \item{L}{labor population}
 \item{M}{area}
 \item{K}{chemical fertilizer}
}}
\usage{
grainproduct
}
\description{
A dataset containing year,Y,L,M,K 5 variables of 27 objects
}
\keyword{datasets}

