% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_happy.R
\docType{data}
\name{happy}
\alias{happy}
\title{Dataset of happiness in chapter14}
\format{a dataframe with 24 rows and 5 variables
\describe{
 \item{happy}{1~10:the degree of happiness,and 10 is the happinest}
 \item{money}{family income,1000$ per unit}
 \item{sex}{lover,1:yes;0:no}
 \item{love}{emotional index,1:lonely;2:stable relationship;3:strong relationship}
 \item{work}{work index,1:no work;3:ok;5:work well}
}}
\usage{
happy
}
\description{
A dataset containing happy,money,sex,love,work of 24 objects
}
\keyword{datasets}

