% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_heart_maxrate.R
\docType{data}
\name{heart_maxrate}
\alias{heart_maxrate}
\title{Dataset of age and maxrate in chapter10}
\format{a dataframe with 2 variables and 15 objects
\describe{
 \item{Age}{Age}
 \item{MaxRate}{maxrate}
}}
\usage{
heart_maxrate
}
\description{
A dataset containing Age and MaxRate of 15objects
}
\keyword{datasets}

