% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_height_weight.R
\docType{data}
\name{height_weight}
\alias{height_weight}
\title{Dataset of students'height and weight in chapter6}
\format{a dataframe with 19 rows and 5 variables
\describe{
 \item{name}{students' name}
 \item{sex}{F:female;M:male}
 \item{age}{students' age}
 \item{height}{students' height}
 \item{weight}{students' weight}
}}
\usage{
height_weight
}
\description{
A data set containing name,sex,age,height and weight 5 variables of 19 objects
}
\keyword{datasets}

