% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importS_GDP.R
\docType{data}
\name{importS_GDP}
\alias{importS_GDP}
\title{Dataset of china's imports demand and total output 1985s-2003s in chapter11}
\format{a dataframe with 19 rows and 3 variables
\describe{
 \item{year}{year}
 \item{actual_imports}{actual_imports}
 \item{actual_GDP}{actual_GDP}
}}
\usage{
importS_GDP
}
\description{
A dataset containing year,actual_imports and actual_GDP 3 variables of 19 objects
}
\keyword{datasets}

