% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_income_pay.R
\docType{data}
\name{income_pay}
\alias{income_pay}
\title{Dataset of income and expenditure in chapter10}
\format{a dataframe with 99 rows and income,expenditure 2 variables
\describe{
 \item{income}{income per month}
 \item{expenditure}{expenditure per month}
}}
\usage{
income_pay
}
\description{
A dataset containing income and expenditure of 99 objects
}
\keyword{datasets}

