% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_income_pay_village.R
\docType{data}
\name{income_pay_village}
\alias{income_pay_village}
\title{Dataset of average income and consumption of village 1985s-2003s in chapter11}
\format{a dataframe with 19 rows and 4 variables
\describe{
 \item{t}{year}
 \item{income}{average pure income per year}
 \item{expend}{average expenditure per year}
 \item{cpi}{consumption price index}
}}
\usage{
income_pay_village
}
\description{
A dataset containing t,income,expend and cpi 4 variables of 19 objects
}
\keyword{datasets}

