% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_investment.R
\docType{data}
\name{investment}
\alias{investment}
\title{Dataset of investment in chapter10}
\format{a dataframe with 19 rows and 4 variables
\describe{
 \item{year}{year}
 \item{total_investment}{total investment current period}
 \item{last_value}{last period stock value}
 \item{last_capital}{last period capital stock}
}}
\usage{
investment
}
\description{
A dataset containing year,total_investment,last_value and last_capital 4 variables of 19 objects
}
\keyword{datasets}

