% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_listed_company.R
\docType{data}
\name{listed_company}
\alias{listed_company}
\title{Dataset of listed company in chapter13}
\format{a dataframe with 30 rows and 8 variables
\describe{
 \item{ARA}{the ratio of receivabls over total asset}
 \item{ASSET}{log of total asset}
 \item{ATO}{turnover rate of asset}
 \item{ROA}{interest rate}
 \item{GROWH}{growth rate of sales revenue}
 \item{LEV}{leverage factor}
 \item{SHARE}{share of the first majority shareholder}
 \item{ST}{1:being treated;0: not}
}}
\usage{
listed_company
}
\description{
A dataset containing ARA,ASSET,ATO,ROA,GROWH,LEV,SHARE and ST 8 variables of 30 objects
}
\keyword{datasets}

