% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_math_stat.R
\docType{data}
\name{math_stat}
\alias{math_stat}
\title{Dataset of students' math and stat score in chapter6}
\format{a dataframe with 24 rows and 2 variables
\describe{
 \item{math}{students' math score}
 \item{stat}{students' stat score}
}}
\usage{
math_stat
}
\description{
A dataset containing math and stat 2 variables of 24 objects
}
\keyword{datasets}

