% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_median_test.R
\name{median_test}
\alias{median_test}
\title{Function of 'median test',and similar to the rank.}
\usage{
median_test(x, median = NA)
}
\arguments{
\item{x}{Numeric vector}

\item{median}{Comparasion object}
}
\value{
The output are the numbers of positive and negative,and P value.
}
\description{
Function of 'median test',and similar to the rank.
}
\examples{
x <- c(21240,4632,22836,5484,5052,5064,6972,7596,14760,15012,18720,9480,4728,67200,52788)
median_test(x, median = 5080)
wilcox.test(x, mu = 5080)

}

