% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_medicion.R
\docType{data}
\name{medicion}
\alias{medicion}
\title{Dataset of population and medicion institutions in sichuan province in 2000s in chapter10}
\format{a dataframe with 16 rows and 3 variables
\describe{
 \item{region}{several regions in sichuan province}
 \item{pop}{population(10 thousand)}
 \item{institutions}{number of medicion institutions}
}}
\usage{
medicion
}
\description{
A dataset containing region,pop and institutions 3 variables of 16 objects
}
\keyword{datasets}

