% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plantarea_outputvalue.R
\docType{data}
\name{plantarea_outputvalue}
\alias{plantarea_outputvalue}
\title{Dataset of agritural plant output and area in chapter11}
\format{a dataframe with 29 rows and 2 variables
\describe{
 \item{plant_area}{area for plant}
 \item{output_value}{output}
}}
\usage{
plantarea_outputvalue
}
\description{
A dataset containing plant_area and output_value 2 variables of 29 objects
}
\keyword{datasets}

