% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plantgrowth.R
\docType{data}
\name{plantgrowth}
\alias{plantgrowth}
\title{Dataset of growth of plant in chapter8}
\format{a dataframe with 30 rows and 2 variables
\describe{
 \item{weight}{weight}
 \item{group}{ctrl,trt1,trt2}
}}
\usage{
plantgrowth
}
\description{
A dataset containing weight and group 2 variables of 30 objects
}
\keyword{datasets}

