% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot3D.R
\name{plot3D}
\alias{plot3D}
\title{Function of 'Plot Three Dimensional Graph'}
\usage{
plot3D(x, y, f)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{f}{A function}
}
\value{
a three dimensional graph
}
\description{
Function of 'Plot Three Dimensional Graph'
}
\examples{
x <- seq(-10, 10, length = 30)
y <- x
f <- function(x, y){r <- sqrt(x ^ 2 + y ^ 2);10 * sin(r) / r}
plot3D(x, y, f)

}

