% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rcmeans}
\alias{rcmeans}
\title{Row and column means}
\usage{
rcmeans(Y)
}
\arguments{
\item{Y}{(matrix) a real-valued matrix.}
}
\value{
A list consisting of the row means (\code{rmean}) and the
column means (\code{cmean}).
}
\description{
Computes the row and column means of a matrix.
}
\details{
An alternative to calculating the degress, \code{rcmeans}
computes the mean of the non-missing entries of each row and
each column of \code{Y}.
}
\examples{
data(comtrade)
rcmeans(comtrade[,,1,1])
rcmeans(comtrade)
}
\author{
Peter Hoff
}

