% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\docType{package}
\name{RDA}
\alias{RDA}
\alias{RDA-package}
\title{Datasets, functions and examples from the book:R Data Analysis-Methods and application.}
\description{
Datasets, functions and examples from the book:R Data Analysis-Methods and application.
}
\section{EDA}{

Function of 'Exploratory Data Analysis'.
}

\section{u_test}{

Function of testing for 'Mean of Norm Distribution' where variance is known.
}

\section{var_test}{

Function of 'Variance Test' for single sample.
}

\section{GQtest}{

Function of 'Goldfeld-Quandt test' for different variance.
}

\section{median_test}{

Function of 'median test',and similar to the rank.
}

\section{proptest}{

Function of 'Proportion Hypothesis Test'.
}

\section{plot3D}{

Function of 'Plot Three Dimensional Graph'.
}

