% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_revenue_tax.R
\docType{data}
\name{revenue_tax}
\alias{revenue_tax}
\title{Dataset of national revenue and tax 1985s-2001s in chapter6}
\format{a dataframe with 12 rows and 3 variables
\describe{
 \item{year}{time}
 \item{revenue}{national revenue(per billion)}
 \item{tax}{national tax}
}}
\usage{
revenue_tax
}
\description{
A dataset containing year revenue and tax 3 variables of 12 objects
}
\keyword{datasets}

