\name{subsample}
\alias{subsample}
\title{Sampling}
\usage{
subsample(x, size, ...)
}
\arguments{
  \item{x}{(vector) A vector of values to be sampled.}

  \item{size}{(integer) The number of values to sample.}

  \item{\ldots}{additional arguments taken by
  \code{sample}.}
}
\value{
A vector.
}
\description{
Subsample from a vector.
}
\details{
This function is essentially the \code{sample} command but
without the bug that sampling a vector \code{x} of length
one returns a permutation of \code{1:x}. This function was
taken from an old help file for \code{sample}.
}
\examples{
sample(1:4)
subsample(1:4)

sample(4)
subsample(4)
}
\author{
Peter Hoff
}

