% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tax.R
\docType{data}
\name{tax}
\alias{tax}
\title{Dataset of national tax revenue 1978s-2012s in chapter10}
\format{a dataframe with 35 rows and 5 variables
\describe{
 \item{time}{time}
 \item{tax}{tax revenue}
 \item{GDP}{GDP}
 \item{expand}{expanditure}
 \item{CPI}{consumption price index}
}}
\usage{
tax
}
\description{
A dataset containing time,tax,GDP,expand and CPI 5 variables of 35 objects
}
\keyword{datasets}

