% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_travel.R
\docType{data}
\name{travel}
\alias{travel}
\title{Dataset of national travel revenue 1994s-2003s in chapter10}
\format{a dataframe with 10 rows and 7 variables
\describe{
 \item{year}{year}
 \item{travelincome}{travelincome}
 \item{population}{travel population}
 \item{income.town}{average travel expenditure in town}
 \item{income_village}{average travel expenditure in village}
 \item{distance_road}{distance on road}
 \item{distance_railway}{distance on railway}
}}
\usage{
travel
}
\description{
A dataset containing year,travelincome,population,income.town... 7 variables of 10 objects
}
\keyword{datasets}

