% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tribes}
\alias{tribes}
\title{Read's highland tribe data}
\format{A sixteen by sixteen by 2 socioarray}
\source{
\url{http://http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm#gama}
}
\description{
Positive and negative relations between 16 New Guinean tribes.
}

