% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_u_test.R
\name{u_test}
\alias{u_test}
\title{Function of testing for 'Mean of Norm Distribution' where variance is known}
\usage{
u_test(data, mu, thegma, alternative = "twoside")
}
\arguments{
\item{data}{Numeric vector}

\item{mu}{comparison object}

\item{thegma}{Numeric value given}

\item{alternative}{twoside or less}
}
\value{
mean(u)value and p value
}
\description{
Function of testing for 'Mean of Norm Distribution' where variance is known
}
\examples{
b <- c(22, 24, 21, 24, 23, 24, 23, 22, 21, 25)
u_test(b, 25, 2.4, alternative = 'twoside')

}

