% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_var_test.R
\name{var_test}
\alias{var_test}
\title{Function of 'Variance Test' for single sample}
\usage{
var_test(x, sigma2)
}
\arguments{
\item{x}{Numeric vector}

\item{sigma2}{Numeric value}
}
\value{
The output are variance,chisq2 value,df and P value
}
\description{
Function of 'Variance Test' for single sample
}
\examples{
x <- rnorm(20, 500, 20)
var_test(x, 400)

}

