% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{YX_nrm}
\alias{YX_nrm}
\title{Normal relational data and covariates.}
\format{A list consisting of a sociomatrix \code{Y} and a matrix of nodal attributes \code{X}}
\usage{
data(YX_nrm)
}
\description{
A synthetic dataset that includes continuous (normal) relational data as
well as information on eight covariates
}
\keyword{datasets}

