\docType{data}
\name{butland_ppi}
\alias{butland_ppi}
\title{Protein-protein interaction data.}
\format{An edgelist for a directed, binary relation}
\usage{
data(butland_ppi)
}
\description{
A network of protein-protein interactions (bindings) from
Butland et al (2005) ``Interaction network containing
conserved and essential protein complexes in Escherichia
coli''. Obtained from
\url{http://pil.phys.uniroma1.it/~gcalda/cosinsite/extra/data/proteins/}.
}
\examples{
data(butland_ppi)
gofstats(el2sm(butland_ppi))
}
\keyword{datasets}

